Upload Src Code files for the incomplete update
These files are source code files for the incomplete v8.0-stable update.

since I am still working on adding the remaining features and fixing the remaining bugs, I will not be providing binaries (or installers for binaries) for this release yet, I will only do that when I finish this release/update completely and become satisfied of how well it performs.

You can compile this release yourself if you want, use some packages like `pyinstaller` (or even better, `auto-py-to-exe` if you prefer using a GUI) and read the respective Wiki page on how to compile it.

But for now, this is the current changelog for this commit of this release:

* Add a keybind to the <kbd>Enter</kbd> button in WiFi ADB connection instructions window to press the **Continue** button.
* Fixed the bug when Temp_Cleaner GUI's Windows.Old cleaner doesn't delete (or even detect) Windows.old folders in System drive.
* Added the Memory dump file cleaner (that file is created whenever a Windows BSOD (Blue Screen of Death) occurs (ntoskrnl shows the collection status as a percentage on the BSOD screen: 100% Complete as a White text on the bottom of the QR Code).
* Added the option to specify a custom full path for the Windows Memory Dump file to use in the cleaning function.
* Added hover shrink effect to the 3 main functionality buttons (the "About", "Settings" and "Start Cleaning" ones).
* Added a new top banner (positioned on top of the main banner, that displays the Temp_Cleaner GUI logo along with a short description) that displays useful information (such as if user needs to clean up temporary files or needs to update their copy of Temp_Cleaner GUI)
* The updater module now no longer spawns a window to show you that you need to update, it will instead change the new informative banner.
* Added the Settings option to set certain thresholds that when passed the new informative banner will prompt you to start cleaning.
* and of course, some minor bug fixes, I won't forget these.